+++ android/src/main/java/io/openmobilemaps/mapscore/map/view/MapView.kt
@@ -19,6 +19,8 @@ import androidx.lifecycle.LifecycleObserver
 import androidx.lifecycle.OnLifecycleEvent
 import androidx.lifecycle.coroutineScope
 import io.openmobilemaps.mapscore.graphics.GlTextureView
+import io.openmobilemaps.mapscore.map.scheduling.AndroidScheduler
+import io.openmobilemaps.mapscore.map.scheduling.AndroidSchedulerCallback
 import io.openmobilemaps.mapscore.shared.graphics.common.Color
 import io.openmobilemaps.mapscore.shared.graphics.common.Vec2F
 import io.openmobilemaps.mapscore.shared.graphics.common.Vec2I
@@ -26,8 +28,6 @@ import io.openmobilemaps.mapscore.shared.map.*
 import io.openmobilemaps.mapscore.shared.map.controls.TouchAction
 import io.openmobilemaps.mapscore.shared.map.controls.TouchEvent
 import io.openmobilemaps.mapscore.shared.map.controls.TouchHandlerInterface
-import io.openmobilemaps.mapscore.map.scheduling.AndroidScheduler
-import io.openmobilemaps.mapscore.map.scheduling.AndroidSchedulerCallback
 import io.openmobilemaps.mapscore.shared.map.scheduling.TaskInterface
 import javax.microedition.khronos.egl.EGLConfig
 import javax.microedition.khronos.opengles.GL10
@@ -35,24 +35,20 @@ import javax.microedition.khronos.opengles.GL10
 open class MapView @JvmOverloads constructor(context: Context, attrs: AttributeSet? = null, defStyleAttr: Int = 0) :
 	GlTextureView(context, attrs, defStyleAttr), GLSurfaceView.Renderer, AndroidSchedulerCallback, LifecycleObserver {
 
-	protected lateinit var mapInterface: MapInterface
+	protected var mapInterface: MapInterface? = null
 		private set
-	protected lateinit var scheduler: AndroidScheduler
+	protected var scheduler: AndroidScheduler? = null
 		private set
 
-	private lateinit var touchHandler: TouchHandlerInterface
+	private var touchHandler: TouchHandlerInterface? = null
 	private var touchDisabled = false
-
-	init {
-		System.loadLibrary("mapscore")
-	}
 
 	open fun setupMap(mapConfig: MapConfig) {
 		val densityExact = resources.displayMetrics.xdpi
 
 		setRenderer(this)
-		scheduler = AndroidScheduler(this)
-		mapInterface = MapInterface.createWithOpenGl(
+		val scheduler = AndroidScheduler(this)
+		val mapInterface = MapInterface.createWithOpenGl(
 			mapConfig,
 			scheduler,
 			densityExact
@@ -62,25 +58,27 @@ open class MapView @JvmOverloads constructor(context: Context, attrs: AttributeS
 				scheduler.launchCoroutine { requestRender() }
 			}
 		})
-		touchHandler = mapInterface.getTouchHandler()
 		mapInterface.setBackgroundColor(Color(1f, 1f, 1f, 1f))
+		touchHandler = mapInterface.getTouchHandler()
+		this.mapInterface = mapInterface
+		this.scheduler = scheduler
 	}
 
 	fun registerLifecycle(lifecycle: Lifecycle) {
-		scheduler.setCoroutineScope(lifecycle.coroutineScope)
+		requireScheduler().setCoroutineScope(lifecycle.coroutineScope)
 		lifecycle.addObserver(this)
 	}
 
 	override fun onSurfaceCreated(gl: GL10?, config: EGLConfig?) {
-		mapInterface.getRenderingContext().onSurfaceCreated()
+		requireMapInterface().getRenderingContext().onSurfaceCreated()
 	}
 
 	override fun onSurfaceChanged(gl: GL10?, width: Int, height: Int) {
-		mapInterface.setViewportSize(Vec2I(width, height))
+		requireMapInterface().setViewportSize(Vec2I(width, height))
 	}
 
 	override fun onDrawFrame(gl: GL10?) {
-		mapInterface.drawFrame()
+		mapInterface?.drawFrame()
 	}
 
 	override fun scheduleOnGlThread(task: TaskInterface) {
@@ -89,22 +87,30 @@ open class MapView @JvmOverloads constructor(context: Context, attrs: AttributeS
 
 	@OnLifecycleEvent(Lifecycle.Event.ON_START)
 	open fun onStart() {
-		scheduler.resume()
+		requireScheduler().resume()
 	}
 
 	@OnLifecycleEvent(Lifecycle.Event.ON_RESUME)
 	open fun onResume() {
-		mapInterface.resume()
+		requireMapInterface().resume()
 	}
 
 	@OnLifecycleEvent(Lifecycle.Event.ON_PAUSE)
 	open fun onPause() {
-		mapInterface.pause()
+		requireMapInterface().pause()
 	}
 
 	@OnLifecycleEvent(Lifecycle.Event.ON_STOP)
 	open fun onStop() {
-		scheduler.pause()
+		requireScheduler().pause()
+	}
+
+	@OnLifecycleEvent(Lifecycle.Event.ON_DESTROY)
+	open fun onDestroy() {
+		setRenderer(null)
+		mapInterface = null
+		scheduler = null
+		touchHandler = null
 	}
 
 	fun setTouchEnabled(enabled: Boolean) {
@@ -128,38 +134,41 @@ open class MapView @JvmOverloads constructor(context: Context, attrs: AttributeS
 			for (i in 0 until event.pointerCount) {
 				pointers.add(Vec2F(event.getX(i), event.getY(i)))
 			}
-			touchHandler.onTouchEvent(TouchEvent(pointers, action))
+			touchHandler?.onTouchEvent(TouchEvent(pointers, action))
 		}
 
 		return true
 	}
 
 	fun setBackgroundColor(color: Color) {
-		mapInterface.setBackgroundColor(color)
+		requireMapInterface().setBackgroundColor(color)
 	}
 
 	open fun addLayer(layer: LayerInterface) {
-		mapInterface.addLayer(layer)
+		requireMapInterface().addLayer(layer)
 	}
 
 	open fun insertLayerAt(layer: LayerInterface, at: Int) {
-		mapInterface.insertLayerAt(layer, at)
+		requireMapInterface().insertLayerAt(layer, at)
 	}
 
 	open fun insertLayerAbove(layer: LayerInterface, above: LayerInterface) {
-		mapInterface.insertLayerAbove(layer, above)
+		requireMapInterface().insertLayerAbove(layer, above)
 	}
 
 	open fun insertLayerBelow(layer: LayerInterface, below: LayerInterface) {
-		mapInterface.insertLayerBelow(layer, below)
+		requireMapInterface().insertLayerBelow(layer, below)
 	}
 
 	open fun removeLayer(layer: LayerInterface) {
-		mapInterface.removeLayer(layer)
+		requireMapInterface().removeLayer(layer)
 	}
 
 	fun getCamera(): MapCamera2dInterface {
-		return mapInterface.getCamera()
+		return requireMapInterface().getCamera()
 	}
+
+	fun requireMapInterface(): MapInterface = mapInterface  ?: throw IllegalStateException("Map is not setup or already destroyed!")
+	fun requireScheduler(): AndroidScheduler = scheduler  ?: throw IllegalStateException("Map is not setup or already destroyed!")
 
 }