+++ shared/src/map/layers/polygon/PolygonLayer.h
@@ -41,6 +41,8 @@ class PolygonLayer : public PolygonLayerInterface,
 
     virtual void addAll(const std::vector<PolygonInfo> & polygons) override;
 
+    virtual void setupPolygon(const std::vector<std::shared_ptr<Polygon2dInterface>> &polygon);
+
     virtual void clear() override;
 
     virtual std::shared_ptr<::LayerInterface> asLayerInterface() override;